Separate command and listener from Main
